requirements.txt
Django==4.2.14
djangorestframework==3.15.2
Markdown==3.6
pillow==10.4.0
sqlparse==0.5.0
typing_extensions==4.12.2
django-filter==24.2
django-dotenv==1.4.2
python-environ==0.4.54
django-ckeditor==6.7.1
django-query-counter==0.4.0
